+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+### 2.1.2 - UNRELEASED
+
+* Hover shows a snapshot of textures and colors
+* Definition provider allows navigation on any index reference
+* View buffer data provided by Accessor (Go to Definition (F12))
+* Changed viewing texture data to match VSCode inline viewer.
+
 ### 2.1.1 - 2017-11-15
 
 * Fixed an issue where glTF commands sometimes wouldn't be available.